move fixtures to associated file
