--- tests/private.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpgIBAAKCAQEAzCgzGvGeAbK7LVYib2EpaNX5m0KWwkAN3s79O2Zr18Eqmowy
-s4TILYKa8s43Dp/H+OZsDTHD7KQDgjKjP/06UUF3q3QbEvWRffRu/akvcXQLLYj9
-mcCxixKu5iQEG3mZ+ZcM0F+9GYXOPegF90qejxzzb6MkK/y5RiUrUDBFfLF9H42T
-9B7/We/B/MlPSES1LZwUeh+WQ1QaZZ43LBOFgfY8wVg1Wczp9ExqEQAGUkNJ4w3V
-0pRAZpURRMDKPCzHzhY80FgRYGBlnxmS/ran/hJoGPomWnqhY/kX6cw4PfyiOnnr
-y4UDsx1TwzQ6D/5Ipq3SL/bIDmR+nXGQ575e1wIDAQABAoIBAQCX1cn0XwA5mRjW
-ZfokRenLO2BxVTvUalkmfWEenVTVWKd724sVskpCdqEQtceCy9iox45S3dEuHOkz
-GsAJJicKKZt721ULa7FyMKmfZAWR5lEt43VgtUwV96Y2KVEYsf6O/wUbsChuY/J4
-XiiZu1hjbY+rVnTfdgzBG4G0GMKFMhN+V+ilqUcgoIqTrpQ1xY0gd3hJnlwBAcZK
-SRNaN+7SiwsvBz9qjXQQ868dRe6rRtG9WW/rc3hcficcKm2+R5vKKT4tzWx5K7Is
-WwsuKrLdt0QkBNoKyPRFS3HYk83yhNohlaeOfaqEAPUoi11Y7gGzsqHP/p7TSRmG
-bi6uWMGhAoGBAOtzFRwYJS1bjQvnK5cTeOQCPBMjELWDnLj7hpLfkxp9nHw6n4cJ
-HIxkG4WfvjxdMVHCX0//ilGFYpANkzbGYxTaJoRPqcsC55HaMkD3JFc9CKHITU/8
-OzaQvc6cLWHuUaYDRdDvoLUURsV4oRrE4xKXBc67rStYHToC9PpXUqwxAoGBAN35
-6bTwoiPrN3Pg7hBWUTmMHk0E8OZqBZG5bPoefDNEYFGdNDuIlYXyGCIFIO5dZKJk
-7w5F8LUhoRPFz4OQJae0KdRwbAAg/yWBIfWBOubsnDP0Uyj87GvA4xBDfZ6BG8C9
-+jlv0gDDR+9x3+Jsagn1pTjMTslcpiwCd4+nl2GHAoGBAIscHo/P0D9hRQ+QUNfd
-AGMu0tG3uQqVrUY6DgO/JsnTB176wABnqL9l8kDb5ekCYr5cqjW41L9foKFd8+m8
-AxRx46nu63Hv0GQDWAe8PXNxsyTrFIq6iz2s/QgSjFm9Fe6RJnHQ8tGqHjpUEFOh
-mkwEgMDpb+IZGIOK9bTnXQgxAoGBAJQJhrZYQRoJVZsfaEVW+6aQbujYobxY+svb
-yPTp/dTHBiGH4PzPM0D4WuXdW7/8Yw/2loODW2vBOMaSBrnsGWDFw5vdOfnAB8lA
-XKB5hgnZxLbVe9x7CUU3h3ciO5XI7vRU/prtmzVNBx2GaVR9Dx8DRyQhBW4/Lsdt
-AcW5dDjjAoGBAMYf5Fupvj0rPp94CIy5GZAFbeex3UuQ5KhIl4WUuJZayiVqXjWr
-tNDhg3MFngq3HUKnfR3h7ArfpbuPklpQxTqZrAHkhXkJxpxin+oxYbjMjEHS/4xd
-x9aLJ3XfeJUsJCpdGzLHx+plZyl87MBkQP5s1DT4n3dugSLkeUYaZx7V
------END RSA PRIVATE KEY-----
