--- tests/pub_cert.crt
@@ -1,21 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDZjCCAk6gAwIBAgIUdOffB8guBa52kkAzpD1ppOFL5nMwDQYJKoZIhvcNAQEL
-BQAwbTELMAkGA1UEBhMCdXMxEDAOBgNVBAgMB3N1YmplY3QxEDAOBgNVBAcMB3N1
-YmplY3QxEDAOBgNVBAoMB3N1YmplY3QxEDAOBgNVBAMMB3N1YmplY3QxFjAUBgkq
-hkiG9w0BCQEWB3N1YmplY3QwHhcNMjAwMTI3MTQ1MjI4WhcNMzAwMTI0MTQ1MjI4
-WjBtMQswCQYDVQQGEwJ1czEQMA4GA1UECAwHc3ViamVjdDEQMA4GA1UEBwwHc3Vi
-amVjdDEQMA4GA1UECgwHc3ViamVjdDEQMA4GA1UEAwwHc3ViamVjdDEWMBQGCSqG
-SIb3DQEJARYHc3ViamVjdDCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEB
-ALgRrv3Oe60PGufwPcXpgxSFPYA+H7XmuJ95v9vsiUxJxRN2YWhi9PqVFIQ2N0qN
-Vm9CjHJD39KNXMM4zndhJHHhUC4PeLzGNqfguahyJcQIUm35u6duL96F1XkCHfdh
-5nxBj25neS5pG4mNtlwBd/bTQKeAaJwTX9VYHUtjEcWiDOPIcLOVt+lYEzHXJhdr
-KfGtx3zGCIsDIDgUaYJDsXxwQ3yBQ3QWQa50g5yWXz6gn2kN8g7FZVGP++ZoF4Rw
-zI6dsMA3AqwNGHUf1/f3vHUOrSooyzPgH+YxAxqYBmLxlOCNcsBX/7Z1cMtH83hS
-M2wIdUk7T5j5lnfCbe5cB/0CAwEAATANBgkqhkiG9w0BAQsFAAOCAQEAE7Uiohrs
-ogFr1bURBdYvVe8L/4+cszNYWBJjs2Bnim0u9YyQ22XxrpUJ4xIsrQIHbwb2v3so
-jyRsFNOR1ZEJwvTX8OmWgLzGVIU878nfjlmlORdvYYf6V9THoT9/sCfVfG/vFyba
-qS/3mfsiDy4oyJGeJYppnzrxaDnQCHbLsqHdL+eKxMtig7sGr3z2DBa42i6aCzXa
-htRE9tQgBWgnXTPOnSCNXp95ImO1wtBw1hkMXvKCs6hE8EhNez4mMlptAGVBEUho
-NrpbCvPtYeexHeg41sM2iYzIAsiHAvp9Z7cMF8ZtIV6MYkiIlV6r4jOSyGn/6zDq
-0B1Ffacjt1UZLg==
------END CERTIFICATE-----
